The default flag was removed from the likelihood_init routine
The policy changed some time ago on whether we should test the input
param against the existing log.param in the folder. Now, we do not do
this anymore, instead, the log.param is always used.

This prevents users from running different cosmological parameters in
a same folder. The test to see whether the folder actually contains a
valid log.param is done in the parser_mp.py file.