Bug fix: behaviour for no varying parameters
when called with no varying parameters, Monte Python should call the
cosmological module and evaluate the likelihood of the desired
experiments for these fixed values. This bug was instead using default
Class values.
Note that this was only affecting the behaviour where every parameters
were fixed, not when only some of them were.