Catching error message when reading from wrong log.param files
A situation where this appeared was after a failed initialization of a
run, the output folder was created, alongside the log.param file.
However, the likelihoods did not yet log their data, so when the code
tries to run from the log.param, it fails because it can not find the
relevant data. This error is now catched, but probably not optimally,
when the log.param is read and that no lines concerning any likelihood
is read. This should help things.