Fixed used of flag "-f 0"
It used to serve to fix values of all the parameters, to create a
fiducial model with desired values, without manually editing the
parameter file. Since the refactoring of the code, this behaviour was
broken, and the code ran without stopping, picking the proper values,
but not realizing it was pointless. It now stops as well. Note that
the old way of doing it still works, but is not recommended.