Implemented a message display function
in the module io_mp, there is now a "message" function, that takes a
string as an argument, as well as a status flag, which can so far be
"error", "warning" or "info". It then formats the string to follow
some ascii-art to represent a warning sign. The change was repercuted
through all files were it made sense.