Starting point of Nested Sampling implementation
added a new flag in the parser, -m, that expects a string. By default,
it is 'MH' for Metropolis Hastings, but can be set to 'NS'.

Then, a file called sampler.py was added, which is not called from
MontePython instead of mcmc. The discrimination between mcmc and
nested sampling will be inside of this new sampler module's function
called `run`. The implementation of the nested sampling will be done
in a separate branch, branching off here