File handling now compatible with Python standards
There used to be many places where a given path was checked for ending '/', and
one was added in case it was missing. This allowed to *add* path together, but
goes against the convention in Python, where path should be joined via os.path,
and no trailing character should be there.