Over sampling implemented
You can now specify an additional field in the parameter file, called
data.over_sampling, which should be a list, starting with a 1, and then any
integer number, to spend more time changing nuisance parameters. It should have
the same length as the number of blocks (so, 1 for the cosmological numbers and
any likelihood with varying nuisance parameters).