Added an option to redefine parameters for analyzing
It is now possible to specify another dictionary in the extra plot file,
as presented in `plot_files/example.plot`. The dictionary is called
`redefine`, and contains redefinitions of parameters.

For instance, we could be interested in computing the posterior for
Omega_m, but only omega_b, omega_cdm and H0 were sampled.

`redefine` would then be:
{'omega_cdm': '(0.01*omega_b+omega_cdm)/(H0/100.)**2'}

and then,

info.to_change = {'omega_cdm': 'Omega_m'}

Note that the names appearing in `redefine` are the old names, before
the renaming. Note also that the scales have to be reintroduced by hand,
as the modification takes place at the level of the raw file.